requirements.txt
pydantic-settings==2.8.4
google-generativeai==0.5.0
groq==0.13.0
requests==2.32.3
openai==1.50.0
python-dotenv==1.0.1
